+++ CHANGES.md
@@ -6,6 +6,8 @@ Features
 * `Structure.getFieldOrder()` supersedes `Structure.setFieldOrder()` and is now required - [@twall](https://github.com/twall).
 * Search ~/Library/Frameworks and /Library/Frameworks on OSX - [@shaneholloway](https://github.com/shaneholloway).
 * Automatic cleanup of native threads (based on suggestions from neil smith) - [@twall](https://github.com/twall).
+* Add android-arm target (thanks to ochafik for initial work).
+* Add `jna.tmpdir` to override temporary JNA storage location - [@twall](https://github.com/twall).
 
 Bug Fixes
 ---------